Move effectively-public notation headers out of `internal`
